requirements.txt
flask
mysql-connector-python
requests
smtplib
ssl
email
flask_bcrypt